removed warnings
git-svn-id: https://deskvox.svn.sourceforge.net/svnroot/deskvox/trunk@1222 04231f92-3938-0410-9931-e931ac552b4f

+++ virvo/virvo/vvrenderer.cpp
@@ -193,7 +193,7 @@ void vvRenderState::setParameter(const ParameterType param, const float newValue
     _useIbr = (newValue != 0.0f);
     break;
   case VV_IBR_MODE:
-    _ibrMode = (vvRenderState::IbrMode)newValue;
+    _ibrMode = (vvRenderState::IbrMode)((int)newValue);
     break;
   default:
     break;
